+++ miner/src/BatchSource.java
@@ -6,6 +6,7 @@ import java.util.ArrayList;
 
 public interface BatchSource
 {
+  public int getSuggestedBatchSize();
   public List<ByteString> readWordsBulk(List<Long> indexes);
 
 }